+++ TODO.md
@@ -4,9 +4,3 @@
 - [ ] CONTRIBUTING.md
 - [ ] schema
 - [ ] remove ts-ignore
-- [ ] fix TODO's
-- [ ] make isomorphic
-- [ ] node crypto : "crypto@1.0.1: This package is no longer supported. It's now
-      a built-in Node module. If you've depended on crypto, you should switch to
-      the one that's built-in."
-- [ ] create example using other common clients, such as `apollo-client`
