+++ package.json
@@ -44,6 +44,7 @@
   },
   "devDependencies": {
     "@types/isomorphic-fetch": "^0.0.36",
+    "@types/json-bigint": "^1.0.1",
     "@types/node": "^18.11.18",
     "esbuild": "^0.17.16",
     "nodemon": "^2.0.20"
